make Context dict-like for compat reason
