+++ parliament/__init__.py
@@ -1,8 +1,8 @@
 # __init__.py
 
+from .invocation import Context
 from . import server
 
-Context = server.Context
 
 __version__ = "0.0.5"
-__all__ = [server, Context]
+__all__ = ["server", "Context"]
